pyproject.toml
[tool.poetry]
name = "hometask-14"
version = "0.1.0"
description = ""
authors = ["Oleksii Petrashov oleksii@petrashov@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.18"
python-multipart = "^0.0.6"
alembic = "^1.11.1"
fastapi = "^0.100.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
fastapi-mail = "^1.3.1"
redis = "^4.6.0"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.33.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
pydantic-settings = "^2.0.2"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["."]
